backend/requirements.txt
asgiref==3.3.4
bcrypt==3.2.0
cffi==1.14.6
Django==3.2.4
django-cors-headers==3.7.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.2
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
six==1.16.0
sqlparse==0.4.1
